about the company
ayuda designs is a company dedicated to offering the best fashion needs to the people around he continent. we deal in the production of raw materials, principally fibres and textiles but also leather and fur; the production of fashion goods by designers, manufacturers, contractors, and others; retail sales; and various forms of advertising and promotion. These levels consist of many separate but interdependent sectors, all of which are devoted to the goal of satisfying consumer demand for apparel under conditions that enable participants in the industry to operate at a profit.