about the proprietor
ayuda designs is a dedicated fashio and design company dealing in the production of raw materials, principally fibres and textiles but also leather and fur; the production of fashion goods by designers, manufacturers, contractors, and others; retail sales; and various forms of advertising and promotion. These levels consist of many separate but interdependent sectors, all of which are devoted to the goal of satisfying consumer demand for apparel under conditions that enable participants in the industry to operate at a profit.